+++ jave-core/src/main/java/ws/schild/jave/filters/ScaleFilter.java
@@ -16,6 +16,26 @@ public class ScaleFilter extends Filter {
   /** 
    * Scale the video to a particular size and maintain aspect ratio.
    * @param toSize What size should the video be scaled to?
+   */
+  public ScaleFilter(VideoSize toSize) {
+    super("scale");
+    addNamedArgument("w", toSize.getWidth().toString());
+    addNamedArgument("h", toSize.getHeight().toString());
+  }
+
+  /** 
+   * Scale the video to a particular size and maintain aspect ratio.
+   * @param scaleExpression What size should the video be scaled to?
+   * Can be an expression like "trunc(iw/2)*2:trunc(ih/2)*2"
+   */
+  public ScaleFilter(String scaleExpression) {
+    super("scale");
+    addOrderedArgument(scaleExpression);
+  }
+  
+  /** 
+   * Scale the video to a particular size and maintain aspect ratio.
+   * @param toSize What size should the video be scaled to?
    * @param foar Should the video be increased or decreased to size?
    */
   public ScaleFilter(VideoSize toSize, ForceOriginalAspectRatio foar) {